requirements.txt
pillow 
flask
numpy 
torchsummary
tqdm
matplotlib 
Albumentations


